alpha 1.0.3

See merge request if3110-2023-k01-01-24/tugas-besar-1-wbd!15

+++ src/Handler/Content/ContentHandler.php
@@ -1,8 +1,10 @@
 <?php
 namespace Handler\Content;
 
+use Core\Application\Services\ContentService;
 use Exception;
 use Handler\BaseHandler;
+use Utils\ArrayMapper\ArrayMapper;
 use Utils\Http\HttpStatusCode;
 use Utils\Response\Response;
 
@@ -10,18 +12,17 @@ use Utils\Response\Response;
 
 class ContentHandler extends BaseHandler {
     
-    protected static $instance;
-    protected $service;
-
-    private function __construct($service)
+    protected static ContentHandler $instance;
+    private function __construct(ContentService $contentService)
     {
-        parent::__construct($service);
+        $this->service = $contentService;
     }
 
-    public static function getInstance($container): ContentHandler {
+    public static function getInstance($contentService): ContentHandler
+    {
         if (!isset(self::$instance)) {
             self::$instance = new static(
-                $container->resolve('contentService')
+                $contentService
             );
         }
         return self::$instance;
@@ -33,30 +34,68 @@ class ContentHandler extends BaseHandler {
      * /api/content?page={p} => get content data at page p
      * /api/content?content_id={id} => get a content with a specific id
      */
-    protected function get($params = null)
+    protected function get($params = null): void
     {
-        if (isset($params['content_id'])) {
-            $content = $this->service->getContentById($params['content_id']);
-            if (is_null($content)) {
-                $response = new Response(true, HttpStatusCode::NOT_FOUND ,"Content(s) not found", null);
-                $response->encode_to_JSON();
-                return;
+        try {
+            $resultArray = [];
+            $page = isset($params['page']) ? intval($params['page']) : 1;
+            $pageSize = isset($params['pageSize']) ? intval($params['pageSize']) : 10;
+
+            if (isset($params['content_id'])) {
+                $content = $this->service->getContentById($params['content_id']);
+                $resultArray[] = $content->toArray();
+            } else {
+                if (isset($params['query']) && isset($params['sortAscending'])) {
+                    $query = $params['query'];
+                    $sortAscending = filter_var($params['sortAscending'], FILTER_VALIDATE_BOOLEAN);
+
+                    $result = $this->service->processContentQuery($query);
+                    $filteredResult = $result;
+
+                    /**
+                     * Filtering part
+                     */
+
+                    if ($sortAscending) {
+                        usort($filteredResult, function ($a, $b) {
+                            return strcmp($a->getTitle(), $b->getTitle());
+                        });
+                    } else {
+                        usort($filteredResult, function ($a, $b) {
+                            return strcmp($b->getTitle(), $a->getTitle());
+                        });
+                    }
+
+                    $totalPages = ceil(count($filteredResult) / $pageSize);
+                    header("X-Total-Pages: " . $totalPages);
+                    $startIndex = ($page - 1) * $pageSize;
+                    $pagedResult = array_slice($filteredResult, $startIndex, $pageSize);
+                } else {
+                    $contents = $this->service->getAllContents();
+                    $totalContents = count($contents);
+                    $totalPages = ceil($totalContents / $pageSize);
+                    header("X-Total-Pages: " . $totalPages);
+                    $page = max(1, min($page, $totalPages));
+
+                    $startIndex = ($page - 1) * $pageSize;
+                    $pagedResult = array_slice($contents, $startIndex, $pageSize);
+                }
+
+                $resultArray = ArrayMapper::mapObjectsToArray($pagedResult);
             }
-            $response = new Response(true, HttpStatusCode::OK ,"Content(s) retrieved successfully", $content->toArray());
+
+            if (!empty($resultArray)) {
+                $response = new Response(true, HttpStatusCode::OK, "data retrieved successfully", $resultArray);
+            } else {
+                $response = new Response(false, HttpStatusCode::OK, "data not found", null);
+            }
+            $response->encode_to_JSON();
+            return;
+        } catch (Exception $e) {
+            $response = new Response(false, HttpStatusCode::NOT_FOUND, "Request failed: " . $e->getMessage(), null);
             $response->encode_to_JSON();
             return;
         }
-
-        $pageNumber = $params['page'] ?? null;
-        $contents = $this->service->getAllContents($pageNumber);
-
-        $contentsArray = [];
-        foreach ($contents as $content) {
-            $contentsArray[] = $content->toArray();
-        }
-
-        $response = new Response(true, HttpStatusCode::OK ,"Content(s) retrieved successfully", $contentsArray);
-        $response->encode_to_JSON();
     }
 
     protected function post($params = null)
@@ -92,7 +131,14 @@ class ContentHandler extends BaseHandler {
             $putData = file_get_contents('php://input');
             parse_str($putData, $_PUT);
 
-            $content_id = $_PUT['content_id'] ?? null;
+            $content_id = $_PUT['content_id'];
+
+            if (is_null($this->service->getContentById($content_id))) {
+                $response = new Response(false, HttpStatusCode::BAD_REQUEST, "Content not found", null);
+                $response->encode_to_JSON();
+                return;
+            }
+
             $title = $_PUT['title'] ?? null;
             $description = $_PUT['description'] ?? null;
             $release_date = $_PUT['release_date'] ?? null;
@@ -125,8 +171,14 @@ class ContentHandler extends BaseHandler {
     protected function delete($params = null)
     {
         try {
+            if (is_null($this->service->getContentById($params['content_id']))) {
+                $response = new Response(false, HttpStatusCode::BAD_REQUEST, "Content not found", null);
+                $response->encode_to_JSON();
+                return;
+            }
             $this->service->removeContent($params['content_id']);
             $response = new Response(true, HttpStatusCode::OK, "Content deleted successfully", null);
+            $response->encode_to_JSON();
         } catch (Exception $e) {
             $response = new Response(false, HttpStatusCode::BAD_REQUEST, "Content deletion failed: " . $e->getMessage(), null);
             $response->encode_to_JSON();
